Added qmake build to the prebuild events. renamed pro file. included QT Bluetooth in the widget configuration in QT VS Tools extension. Renamed NippleBluetooth to BluetoothClientViewModel in conjunction with naming conventions. Seperated them into .h and .cpp files. closes #33
