+++ Wilbur/project.pro
@@ -1,0 +1,1 @@
+QT += bluetooth
