+++ Wilbur/prototypeControllerView.cpp
@@ -66,7 +66,6 @@ void PrototypeControllerView::executeControl(buttonType button, int newValue)
     std::string objectName = "N/A";
     bool actionSuccess = false;
 
-
     switch (button)
     {
         case PUMP:
