--- Wilbur/test.pro
@@ -1,1 +1,0 @@
-QT += Bluetooth
